LANGUAGE USED:
In our project we are going to using the basic as well as the object-oriented Programming (OOP) techniques of C++ as well as an external library of C++ know as Simple and Fast Multimedia library (SFML). In our project we used different techniques from SFML and C++ to create a small game on window rather than on the console of visual studio. we created animation that move and get destroyed on a certain condition and the player gain some point whenever the player kill/destroy the enemy animation.